Admin SignOut Api added

+++ Flipkart Clone/admin/src/actions/auth.action.js
@@ -56,9 +56,22 @@ export const isUserLoggedIn = ()=> {
 
 export const signout = () => {
   return async dispatch => {
-    localStorage.clear();
+
+    dispatch({type:authConstants.LOGOUT_REQUEST});
+
+    const res = await axios.post('admin/signOut');
+
+    if(res.status===200){
+      localStorage.clear();
     dispatch({
-      type: authConstants.LOGOUT_REQUEST
+      type: authConstants.LOGOUT_SUCCESS
     });
+    }
+    else{
+      dispatch({
+        type:authConstants.LOGOUT_FAILURE,
+        payload:{error:res.data.error}
+      })
+    }
   }
 }
